Move `stop` package within .gtfs package
