+++ domain/src/test/java/org/mobilitydata/gtfsvalidator/domain/entity/gtfs/stops/GenericNodeTest.java
@@ -20,7 +20,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mobilitydata.gtfsvalidator.domain.entity.gtfs.EntityBuildResult;
 import org.mobilitydata.gtfsvalidator.domain.entity.notice.error.MissingRequiredValueNotice;
-import org.mobilitydata.gtfsvalidator.domain.entity.stops.GenericNode;
 
 import java.util.ArrayList;
 import java.util.Collections;
